airflow/requirements.txt
great-expectations==0.18.22
psycopg2-binary
pyyaml
apache-airflow
coverage
pytest
pytest-cov
openpyxl>=3.1.2